move functions to lazy-load amCharts into testable utils module
