+++ app/utils/lazy-loading.js
@@ -1,0 +1,3 @@
+export {
+  loadAmChartsFiles
+} from 'ember-cli-cedar/utils/lazy-loading';
